move `R_DrawBackground` to r_draw.c, remove old code
