+++ src/m_menu.c
@@ -2025,63 +2025,6 @@ menu_t CompatDef =                                           // killough 10/98
 //
 // killough 11/98: rewritten to support hires
 
-void R_DrawBackground(char *patchname, byte *back_dest)
-{
-  int x,y;
-  byte *back_src, *src;
-
-  V_MarkRect (0, 0, SCREENWIDTH, SCREENHEIGHT);
-
-  src = back_src = 
-    W_CacheLumpNum(firstflat+R_FlatNumForName(patchname),PU_CACHE);
-
-  if (hires)       // killough 11/98: hires support
-#if 0              // this tiles it in hires
-    for (y = 0 ; y < SCREENHEIGHT*2 ; src = ((++y & 63)<<6) + back_src)
-      for (x = 0 ; x < SCREENWIDTH*2/64 ; x++)
-	{
-	  memcpy (back_dest,back_src+((y & 63)<<6),64);
-	  back_dest += 64;
-	}
-#else              // while this pixel-doubles it
-/*
-      for (y = 0 ; y < SCREENHEIGHT ; src = ((++y & 63)<<6) + back_src,
-	     back_dest += SCREENWIDTH*2)
-	for (x = 0 ; x < SCREENWIDTH/64 ; x++)
-	  {
-	    int i = 63;
-	    do
-	      back_dest[i*2] = back_dest[i*2+SCREENWIDTH*2] =
-		back_dest[i*2+1] = back_dest[i*2+SCREENWIDTH*2+1] = src[i];
-	    while (--i>=0);
-	    back_dest += 128;
-	  }
-*/
-    for (y = 0; y < SCREENHEIGHT<<1; y++)
-      for (x = 0; x < SCREENWIDTH<<1; x += 2)
-      {
-          const byte dot = src[(((y>>1)&63)<<6) + ((x>>1)&63)];
-
-          *back_dest++ = dot;
-          *back_dest++ = dot;
-      }
-#endif
-  else
-/*
-    for (y = 0 ; y < SCREENHEIGHT ; src = ((++y & 63)<<6) + back_src)
-      for (x = 0 ; x < SCREENWIDTH/64 ; x++)
-	{
-	  memcpy (back_dest,back_src+((y & 63)<<6),64);
-	  back_dest += 64;
-	}
-*/
-    for (y = 0; y < SCREENHEIGHT; y++)
-      for (x = 0; x < SCREENWIDTH; x++)
-      {
-        *back_dest++ = src[((y&63)<<6) + (x&63)];
-      }
-}
-
 void M_DrawBackground(char *patchname, byte *back_dest)
 {
   if (setup_active && menu_background)
